AIModel/requirements.txt
pandas
numpy
scikit-learn
tensorflow==2.11.0
tensorflow-estimator
keras-tuner
Pillow
tqdm